Fix faces dtype: cast to int32 before to_glb (cumesh expects Int not Long)
torch.unique returns int64 (Long); cumesh.init() requires int32 (Int).
Cast the remapped face indices to torch.int32 before passing to o_voxel.

+++ engines/trellis_v2.py
@@ -204,7 +204,7 @@ class TRELLIS2Engine(Engine):
                 f"({raw_mesh.vertices.shape[0]:,} → {unique_idx.shape[0]:,})"
             )
             vertices = raw_mesh.vertices[unique_idx]
-            faces = inverse.reshape(raw_mesh.faces.shape)
+            faces = inverse.reshape(raw_mesh.faces.shape).to(torch.int32)
         else:
             vertices = raw_mesh.vertices
             faces = raw_mesh.faces
